server/requirements.txt
fastapi==0.108.0
groq==0.20.0
pyannote-audio==4.0.4
silero-vad==6.2.1
torchaudio==2.10.0
uvicorn==0.25.0
websockets>=12.0
pydub>=0.25.1
python-dotenv>=1.0.0
scipy>=1.11.0